V8: 重构鸭翼状态机 — 删除升降舵极限偏转，新增接地收回阶段
鸭翼控制逻辑修改：
- 删除两处着陆阶段的升降舵强制最大低头偏转
- 新增 FW_CANARD_RETR 参数：接地后鸭翼收回目标角度（默认0.5/中立）
- 阶段1接地后立即收回到可配置角度（替代原来的保持偏角）
- FW_CANARD_BRKD 默认值改为5.0秒（刹车延迟）
- 修复 ActuatorEffectivenessControlSurfaces constrain 类型编译错误
- 空气frame参数按字母顺序重排并翻译中文注释

+++ src/modules/control_allocator/VehicleActuatorEffectiveness/ActuatorEffectivenessControlSurfaces.cpp
@@ -81,7 +81,7 @@ void ActuatorEffectivenessControlSurfaces::updateParams()
 	int32_t count = 0;
 
 	if (param_get(_count_handle, &count) == 0) {
-		_count = math::constrain(count, 0, MAX_COUNT);
+		_count = math::constrain((int)count, 0, MAX_COUNT);
 	}
 
 	for (int i = 0; i < _count; i++) {
